+++ docs/dummy_settings.py
@@ -1,3 +1,0 @@
-# Dummy Django settings file required to import 
-# crispy-form-foundation when building the documentation
-SECRET_KEY = 'dummy-key'