+++ manage.py
@@ -1,12 +1,12 @@
 #!/usr/bin/env python
 """
-Sandbox management script
+Django management script for sandbox
 """
 import os
 
 from django.core import management
 
 if __name__ == "__main__":
-    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "settings.demo")
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "sandbox.settings.demo")
 
     management.execute_from_command_line()